feat(deployment): added terraform and https
* feat(deployment): added terraform

We used azure as backend. It deploy the entire infra need to run pipelines. The containers are deployed. It manages the dns records in ovh as well.

* fix: uses caddy

* bump: version 0.4.1 → 0.5.0

* feat(deployment): set exposed ports

In order to use https, we exposed only 80 and 443 from the container group

* fix(web): use https and remove port from url

We are now using https and a reverse proxy so we no longer need to specify the port

* feat(deployment): uses terraform in workflow

* ci: used redirection after tf plan