refactor(model): refactored package to work properly
* fix(deployment): disable soft delete

When destroying a workspace on azure by default, it is soft deleted and we don't want that

* refactor(pipeline): refactored pipeline code

We now have a docker image containing all the dependencies that is used in a dedicated azureml environment.

* refactor(model): working model package

We refactored the code in order to be able to use the package correctly

* refactor(api): uses model package

We are now using the model package. We had to modify the dockerfile to be able to include that model into the image.

* refactor(ci): working ml-pipeline

We had to modify the working directory of the steps

* fix(ci): changed working directories to adapt to new structure